Phase 4 All tests passing, need to write JUnit tests for DAO methods

+++ chess/src/main/server/services/LogoutService.java
@@ -1,6 +1,7 @@
 package server.services;
 
 import server.DAO.AuthTokenDAO;
+import server.MyServerException;
 import server.requests.AuthTokenRequest;
 
 /**
@@ -13,6 +14,9 @@ public class LogoutService {
      */
     public void logout(AuthTokenRequest req) {
         AuthTokenDAO authTokenDAO = AuthTokenDAO.getInstance();
+        if (authTokenDAO.find(req.getAuthToken()) == null) {
+            throw new MyServerException("Unauthorized", 401);
+        }
         authTokenDAO.remove(req.getAuthToken());
     }
 }
